scm/azdo/requirements.txt
requests>=2.31.0
rich>=13.7.0
pandas>=2.1.0
openpyxl>=3.1.2
matplotlib>=3.8.0
urllib3>=2.0.0
python-dotenv>=1.0.0
